Convert routes tab to stack screen
